refactor: utils 함수 폴더구조 개선
