+++ src/modules/auth/auth.service.ts
@@ -5,7 +5,7 @@ import { Repository } from 'typeorm';
 import { OAuthApp } from 'octokit';
 import { configService } from '@/config/config.service';
 import { BadRequestException } from '@nestjs/common';
-import { tokenService } from '@/utils/token.service';
+import { tokenService } from '@/utils';
 
 type UserInfo = {
   snsId: number;
