+++ src/utils/judge/index.spec.ts
@@ -1,5 +1,5 @@
 import { v4 as uuidv4 } from 'uuid';
-import { judge } from './judge';
+import { judge } from '.';
 
 describe('judge util test', () => {
   const testCaseTemplate = `
