+++ src/utils/template/index.ts
@@ -1,0 +1,6 @@
+export const responseTemplate = (message: string, data: unknown) => {
+  return {
+    message,
+    data,
+  };
+};
